# test/sensorapi-swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"SensorAPI"},"paths":{"/api/Nodes":{"get":{"tags":["Nodes"],"operationId":"GetNodes","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Node"}}}}},"post":{"tags":["Nodes"],"operationId":"PostNode","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"node","in":"body","required":false,"schema":{"$ref":"#/definitions/Node"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Node"}}}}},"/api/Nodes/{id}":{"get":{"tags":["Nodes"],"operationId":"GetNode","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Node"}}}},"put":{"tags":["Nodes"],"operationId":"PutNode","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"},{"name":"update","in":"body","required":false,"schema":{"$ref":"#/definitions/Node"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Node"}}}},"delete":{"tags":["Nodes"],"operationId":"DeleteNode","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Node"}}}}},"/api/Sensors":{"get":{"tags":["Sensors"],"operationId":"GetSensors","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Sensor"}}}}},"post":{"tags":["Sensors"],"operationId":"PostSensor","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"sensor","in":"body","required":false,"schema":{"$ref":"#/definitions/Sensor"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Sensor"}}}}},"/api/Sensors/{id}":{"get":{"tags":["Sensors"],"operationId":"GetSensor","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Sensor"}}}},"put":{"tags":["Sensors"],"operationId":"PutSensor","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"},{"name":"update","in":"body","required":false,"schema":{"$ref":"#/definitions/Sensor"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Sensor"}}}},"delete":{"tags":["Sensors"],"operationId":"DeleteSensor","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Sensor"}}}}},"/api/Values":{"get":{"tags":["Values"],"operationId":"GetValues","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"},{"name":"ts","in":"query","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/SensorValue"}}}}},"post":{"tags":["Values"],"operationId":"PostValue","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/SensorValue"}}],"responses":{"200":{"description":"Success"}}}},"/api/Values/{id}":{"get":{"tags":["Values"],"operationId":"GetSensorValues","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int64"},{"name":"limit","in":"query","required":false,"type":"integer","format":"int32"},{"name":"ts","in":"query","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/SensorValue"}}}}}}},"definitions":{"Node":{"type":"object","properties":{"id":{"format":"int64","type":"integer"},"name":{"type":"string"},"description":{"type":"string"}}},"Sensor":{"type":"object","properties":{"sensorId":{"format":"int64","type":"integer"},"nodeId":{"format":"int64","type":"integer"},"name":{"type":"string"},"type":{"format":"int32","type":"integer"},"description":{"type":"string"}}},"SensorValue":{"type":"object","properties":{"id":{"format":"int64","type":"integer"},"sensorId":{"format":"int64","type":"integer"},"timestamp":{"format":"int64","type":"integer"},"value":{"type":"string"}}}}}